+++ vite.config.ts
@@ -1,6 +1,5 @@
 import { defineConfig } from 'vite';
 import preact from '@preact/preset-vite';
-import { viteStaticCopy } from 'vite-plugin-static-copy'
 
 
 // https://vitejs.dev/config/
@@ -8,13 +7,5 @@ export default defineConfig({
 	base: "https://rwnx.github.io/scrunked/",
 	plugins: [
     preact(),
-    viteStaticCopy({
-      targets: [
-        {
-          src: 'CHANGELOG.md',
-          dest: ''
-        }
-      ]
-    })
   ],
 });